Added ability to copy the config default folder to user home
It skips if it is found. (dont overwrite user defined config.yaml)

+++ install.sh
@@ -1,19 +1,33 @@
 #!/bin/bash
 
+CONFIG_FOLDER="/home/$USER/.pipymo"
 
 SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 
-sudo rm -f /usr/local/bin/pipymo
+if [ -d "$CONFIG_FOLDER" ]; then
+    echo ".pipymo folder exists, skipping copy of defaults..." 
+else
+    mkdir "$CONFIG_FOLDER"
+    cp "$SCRIPT_DIR/config/*" "$CONFIG_FOLDER/"
+fi
 
-sudo chmod 744 $SCRIPT_DIR/__main__.py
-sudo chmod +x  $SCRIPT_DIR/__main__.py
+if [ -f "/usr/local/bin/pipymo" ]; then
+    echo ".pipymo was previously installed removing /usr/local/bin/pipymo" 
+    echo "if you are running it as a script please restart the flask server using:"
+    echo "-> sudo systemctl restart pipymo"
+    sudo rm -f /usr/local/bin/pipymo
+fi
 
-sudo ln -s $SCRIPT_DIR/__main__.py /usr/local/bin/pipymo
+sudo chmod 744 "$SCRIPT_DIR/__main__.py"
+sudo chmod +x  "$SCRIPT_DIR/__main__.py"
 
-sudo cp $SCRIPT_DIR/pipymo.service /etc/systemd/system/
+sudo ln -s "$SCRIPT_DIR/__main__.py" "/usr/local/bin/pipymo"
 
-pip install -r $SCRIPT_DIR/requirements.txt
+sudo cp "$SCRIPT_DIR/pipymo.service" "/etc/systemd/system/"
+
+pip install -r "$SCRIPT_DIR/requirements.txt"
 
 
-echo "if you want enable the http service with\n sudo systemctl enable pipymo && sudo systemctl start pipymo"
+echo "if you want enable the http service with:"
+echo "sudo systemctl enable pipymo && sudo systemctl start pipymo"
 echo "Done! use it with -> pipymo echo"
